implement stream reponse version
- introduce more modular approach for sub components
- there is an error encountered receiving response
from the model such that some responses are being read
from stdout malformed(cut from the half, not complete
json). So, added a logic to store the malformed reponses
cumulatively and try json decode each time. Assumption is
the missing part of the json is alwasy sent alone in the next
line.
- beautify buf window, and set ftype to markdown.